Initial working application.
User inputs an ordered sequence of actions; these are exercises and breaks.
For each of these, he specifies its duration.
The number of sets specifies how many time the whole sequence is repeated.

When ran, the actions are converted to events.
The event are then executed sequentially.
The execution of the events can be cancelled.

Currently, actions are converted to sound events.