Use an expression to specify workouts of a round
It allows faster inputting of workouts,
compared to a data grid method.

The expression is a sequence of `{number} E|B` sub-expressions,
separated by whitespace.
Each subexpression denotes a single exercise ('E`) or a break (`B`),
with a specific number of seconds (`{number}`).

The UI parses the expression up to its end,
or the first place where parse fails.
It also displays the parsed expression,
so it is possible to wheter the inputted expression
represents the expected workout plan.